Add simple d3 example
This commit adds a simple d3 example to the project. I made minor
updates so that the styles would render properly, so CSS modules was
stripped out for the moment. I also added the Extract Text Plugin, which
we will need at some point but now now. To learn how to do what I did with the d3
data visit:
https://egghead.io/lessons/d3-get-started-with-d3?course=introduction-to-d3