dd redux and react router to project
Added redux and react router to project. Instead of the component
getting data from state it will now get it from props. This is just the
first implementation so it's just very bare bones right now. I took from
two projects to put this together:
https://github.com/reactjs/redux/tree/master/examples/real-world
https://github.com/reactjs/redux/tree/master/examples/async

Added babel-root-import so that I didn't have to remember how many
levels up a component was. Check this repo for more info:
https://github.com/michaelzoidl/babel-root-import

Added Ramda library to help us transform data because it's functional.
Yeah, FP baby! Check it out here:
http://ramdajs.com/0.22.1/#

Added selectors so that data is memoized. Check up on selectors here:
https://github.com/reactjs/reselect

This was all done in preparation for the app to be
progressive/offline-first. Yay!