Update components and containers
Update of the components and containers. I originally was going to have
the App component fetch all data for the projects but I thought it might
be better for each individual project/container do that itself. I also
changed the components to be stateless functional components in, since
they hold no state.