Update defaultProps for BarChart component
Soooooooooo...in React there is the idea of propTypes and defaultProps.
The defaultProps are kind of a placeholder for when the props are
undefined, to handle undefined errors. This allows the component to
wait, and not throw and error, until the props exist and populate the component.
It's more or less a safeguard for this. I did not have an issue with
this initially because I did not check using any more values than the
initial object it self in the component. When I then needed to populate
the component with other values then the errors appeared. Value is
undefined, for example. Therefore I looked for how to solve this.
Soooooooooo... to make this a bit easier to use and manage, I changed
the default value to be an object and used the Ramda.merge function to
merge the data together, considering that the object we are using is
nested with two objects that both point to a prop named data. So when it
merges it makes just one object with a data field. Because Ramda
function are all curried, when the initial data value does not
exist, that's okay, because the function then will not execute. When the
value, is available, it will execute, and therefore will not throw an
error on the initial render. This now helps us to render the BarChart
component without any errors. Yay!