Updated webpack to point to your project
Sooooooo... I didn't mention that I changed the entry point of the app
from being main.js to index.js. I changed it for your branch so that you
can see what you are working on. If you want to see what it would look
like just update the entry point to be index.js and you will get the
ugly view of a horrible and incomplete UI experience. Sorry, not sorry!