Update hoc stuff
Updated the higher order component module to just include the
windowDimensions function since that is the only data set we have that
deals with state. If other hocs are needed I will add them. I removed
the composed library because I thought that was a little overkill in
regards to having a separate folder just for that.