Add fetchData hoc
So I got a little carried away. I wanted to see how I could get away
with not using the componentDidMount lifecycle method and not having to
depend on redux to manage my state while also learning how to use
localStorage. Well I found that out by creating a fetchData hoc that
will fetch the data for the component. What I want to end up doing is
using a different way of storing app data, the next will be
sessionStorage, and after that I want to try and use redux-observable.
This is just one way of doing things and I want to expose myself, and
you @jordan-carney, to all the wonderful ways of doing stuff. Yeah!