Add container for HeatMap chart
Added a container for the Heat Map chart. This combines all of what I
have done previously and builds off of it somewhat. Two things to keep
in mind, (1) I am using pattern matching from react-router, and (2) I
will only be using pattern matching for this component and not any
future or previous components. It is a bit of a hack that I will change
for the next component container I create but by no means is this best
practice. Pattern matching works best for templating, which we are not
doing. It did this more so as an experiment to see if I could render
Components dynamically, which I don't believe we can do at the moment.
Secondly, I did this just to play around with pattern matching and have
it as part of our project. It's just me experimenting and by no means is
it something that I think would benefit a production level app. Again,
just experimentation and fun, nothing else. I just wanted to see how far
I could take React and recompose.