Feature: Add All Graphs (#43)
* Added ScatterPlot Graph from Free Code Camp
Made App the parent component for each graph of the Free Code Camp D3 project track (so far bar-graph, and scatter-plot - more to come)
Added Mouse-out handlers to compontents to tooltips don't hang
Minor adjustment and additions to styles

* trivial update

* Updated webpack to point to your project

Sooooooo... I didn't mention that I changed the entry point of the app
from being main.js to index.js. I changed it for your branch so that you
can see what you are working on. If you want to see what it would look
like just update the entry point to be index.js and you will get the
ugly view of a horrible and incomplete UI experience. Sorry, not sorry!

* added outline for heatmap component

* added FCC Heat Map project to graphs

* added Force Directed Country graph

* meteor map initial commit

* added acceptable responsive styling to the meteor-map